requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# c:\e\pyworkspace\pingsome\pingsome.py: 2
PyYAML == 6.0

# c:\e\pyworkspace\pingsome\.ipynb_checkpoints\showping-checkpoint.ipynb: 2
# c:\e\pyworkspace\pingsome\showping.ipynb: 2
matplotlib == 3.6.2

# c:\e\pyworkspace\pingsome\.ipynb_checkpoints\showping-checkpoint.ipynb: 1
# c:\e\pyworkspace\pingsome\pingsome.py: 5
# c:\e\pyworkspace\pingsome\showping.ipynb: 1
pandas == 1.5.2

# c:\e\pyworkspace\pingsome\pingsome.py: 3
ping3 == 4.0.4

# c:\e\pyworkspace\pingsome\.ipynb_checkpoints\showping-checkpoint.ipynb: 3
# c:\e\pyworkspace\pingsome\showping.ipynb: 3
tables == 3.8.0
